+++ src/CrudServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Backpack\CRUD;
 
 use Route;
+use Prologue\Alerts\Facades\Alert;
 use Illuminate\Support\ServiceProvider;
 
 class CrudServiceProvider extends ServiceProvider
@@ -56,6 +57,8 @@ class CrudServiceProvider extends ServiceProvider
                     '--tag' => 'public',
                 ]);
             }
+        } else {
+            $this->checkLicenseCodeExists();
         }
 
         // use the vendor configuration file as fallback
@@ -111,5 +114,18 @@ class CrudServiceProvider extends ServiceProvider
         }
 
         return false;
+    }
+
+    /**
+     * Check to to see if a license code exists.
+     * If it does not, throw a notification bubble.
+     *
+     * @return void
+     */
+    private function checkLicenseCodeExists()
+    {
+        if (! env('BACKPACK_LICENSE')) {
+            Alert::add('warning', "<strong>You're using unlicensed software.</strong> Please <a target='_blank' href='http://backpackforlaravel.com'>purchase a license code</a> to hide this message.");
+        }
     }
 }